Allow multiple tids to use resourceful at a time with each getting its own resource consumption.
This is implement as a linked list in the kernel that stores the list of pids that have called acct_next.

When should_acct is called, we linearly scan through this list, seeing if there is an entry with the relevant tid.

This is pretty inefficient...every acct_next now has a kmalloc.  Moreover we do a linear walk over a data structure (although the size is expected to be small). Moreover, we do not access memory serially, so it will be slow.

There is no concurrency protection (yet).