Initial implementation of concurrency support for the stap shim.
The shim now kmallocs a struct accounting each time that acct_next is
called, and returns a pointer to this when should_acct returns true.

The systemtap script then keeps track of which struct accounting it
should be writing to at the moment.

We will soon not kmalloc a struct accounting for each acct_next, rather
reusing them when the res module has finished using that particular
element from the relay channel.