Initial implementation of tracking an ID each time that a acct_next call is made that is propogated throughout the lifetime of the syscall.
This will be built upon so that we can have a bunch of syscalls in
flight and track resource usage of each of them indepdently.

At present we return -1 from read_acct_data if multiple syscalls are
accounted for simultaneously, and the one currently at the head of
relayfs is not the one we are looking for. This will be fixed later.
Moreover, we also need to be able to return struct accountings to the
free_accounting_pool after we have memcpy'd them out of the relayfs, to
reduce our use of kmalloc/kfree.

This patch has required some changed to our user-facing
api. Specifically we have to carry around a slightly more rich
datastructure than a char *, as a rsfcfl_handle. This must be passed to
acct_next so we can identify a counter for a syscall.

Throughout the kernel side we track syscalls using the pair (pid,
id) where id is a per-process monotonically-increasing counter
(that can overflow). We attach this to a struct accounting so we can
compare that the result from the relayfs matches the syscall we expected
to make.