Attempt to create a static system for finding subsystems to complement the dynamic one.
This script uses the cscope database to statically determine call-sites
for each non-static function in the linux kernel.

It is unfortunately very slow at the moment (several hours to run),
but could be sped up. The nice thing is that this only has to be done
once per kernel, rather than repeated on different ``representative
workloads'''.