Replace relay with a char driver.
This let us write to the page from userspace, which relay doesn't.

This commit is a regression in that it effectively does a relay_reset,
however I shall fix that.

This new scheme will be cool as we can map memory for each process and
so have to do a lot less searching (practically none!) to find where
we're accounting