Move Initialisation of rscfl into a probes.{h,c}.
The idea is to gradually move away from systemtap. It gives us little value and we have to keep HACKing stuff to embed C, and have wrappers around every function. We can also only pass longs. Seems better to just insert the probes manually as this is all systemtap is doing for us