clang-format applied to entire project.
Our slapdash style had become annoying. This patch uses clang-format to give it a consistent style.

I knocked up a .clang-format file that more-or-less follows our usual style. Please run it before you commit.

The major change is that this removes spaces after casts. The option to stop this was added in clang-format 3.7, which isn't in trusty