Major refactor of subsys_static_discovery to use MAINTAINERS file.
Previously we identified a subsystem as a top-level directory, but this
is hugely course, and generally a bad idea.  This patch defines a
subsystem by using the subsystems defined in the MAINTAINERS file. This
gives us a much more fine-grained notion of a subsystem. Moreover, any
large repo should have a similar file, so our approach can be adapted to
another operating system.

The script now finds all callq functions by scanning objdump. For each
function, it takes the addresses, and pumps them through addr2line to
find which file they are in. I then call get_maintainers.pl, which
identifies the subsystem they're in from the MAINTAINERS file.