Store per-subsystem accounting details in a sparse data structure.
This patch breaks up our struct accounting so that each subsystem stores
resources per subsystem.

This breaks up our memory=mapped region into an area for struct
accounting, and another area for struct subsys_accounting (which stores
the resources for a specific subsystem). The struct accounting then
stores indices into this second area. We do this because we want to be
able to equally address struct subsys accountings, no matter where the
pages are mapped.

In order to support this patch we have to increase the size of memory
mapped shared pages, as we are now storing information regarding more
subsystems than we used to.