This patch is a HACK to stop us accounting only for the netlink subsystem.
Previously we have used netlink to add an expression of
interest. However, netlink itself crosses function boundaries within
NETWORKINGGENERAL. This causes resoucrseful to look at the interest
list, and see that it should be accounting a syscall for the current
pid! We therefore start measuring values.

When the function in interest returns, we notice that our current
nesting is back at 0, so call acct_clear. This has the effect of not
accounting the next syscall.

This patch is a HACK. It assumes that when calling netlink we go down
nestings once. We therefore use a flag to say that we don't take
measurements when we first increase nesting, but rather wait until we
decrease nesting to 0 and then start increasing it again.

This seems to work, at least for the demo. We might want to think a bit
harder about it later on.