This patch removes netlink, and replaces it with shared memory.
Netlink was a pain. It used a bunch of systemcalls, which meant we had
to use a HACK to avoid accounting for netlink.

This patch removes netlink, and replaces it with shared mmaped memory,
so raising an interest does not involve a system call. The
implementation is based on having a data driver that shares struct
accounting and struct subsys_accounting, and a ctrl page that lets
programs register resourceful interests. Both pages are per-pid, which
has the nice benefit that we no longer have to walk a list of interests
for the system.

This turned out to actually be a fairly sized patch. I had to do a bit
of refactoring as our previous solution relied on us getting a kernel
execution context.