Add value validation to fix a segfault.
This is a fix for #31.

If we cannot find a free subsystem we return NULL in the get_subsys
function. When entering a subsystem we treated this properly, by
assuming that we should not account for the subsystem. However on
subsystem exit, we assume that we have a corresponding start, which is
not true if we cannot allocate memory for the subsystem, but want to
account for it.

This patch prints to KERN_ERR and doesn't try to dereference the NULL
when we cannot allocate memory for subsystems.