Add API and data layout versioning.
At the API level, check that the major version of the rscfl header used by an application is
the same as the one of the library we're dynamically linking to.
Whether this results in a warning or a full-blown error is controlled by defining the symbol
RSCFL_ERR_VER_MISMATCH

At the kernel level, add a data layout version as the first value in the control character device.
The API then checks when doing the mmap-ing that the data layout versions are the same between the kernel and the user-space library.

This patch does not introduce any incompatible API changes.