Avoid clobbering the return value of system calls when one system call executes another.
If a kernel fn executes sys_y, then without this patch sys_y's return
address is modified to be the kernel function that called it. This then
makes other threads executing the same system call return to the wrong
place.

This patch modifies such callsites so that they skip the beginning of
all SyS_ functions, thereby not hitting the kamprobes trampoline.

Conflicts:
	src/kernel/tapsets/kamprobes.c